// frontend/src/generated/models/ModelInfo.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ModelInfo = {
    name?: string | null;
    repo?: string | null;
    file?: string | null;
    contextSize?: number | null;
    loadedAt?: string | null;
};